Part2-AzureFunctions_Deployment/requirements.txt
azure-functions
beautifulsoup4==4.8.2
blis==0.4.1
Bottleneck==1.3.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
fastai==1.0.60
fastprogress==0.2.2
idna==2.8
importlib-metadata==1.5.0
kiwisolver==1.1.0
matplotlib==3.1.3
murmurhash==1.0.2
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
packaging==20.1
pandas==1.0.1
Pillow==7.0.0
plac==1.1.3
preshed==3.0.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
scipy==1.4.1
six==1.14.0
soupsieve==1.9.5
spacy==2.2.3
srsly==1.0.1
thinc==7.3.1
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
urllib3==1.25.8
wasabi==0.6.0
zipp==2.2.0
